Standardize server middleware
