+++ src/server.ts
@@ -1,24 +1,17 @@
 import { Hono } from "hono";
 import { handle } from "hono/cloudflare-pages";
-import { APIError } from "./core/api-error";
-import BaseModelClass from "./core/base.modelclass";
-import { Env } from "./types/api.types";
-import { ServerConfig, defaultCors } from "./types/server.types";
+import type { Env } from "./types/api.types";
+import type { ServerConfig } from "./types/server.types";
+import { bindDatabase, defaultCors, defaultErrorHandler } from "./utils";
 
 export class BackendServer {
   private _app: Hono<Env, any, any>;
-  private _config: ServerConfig = {
-    name: "default"
-  };
   public configured: boolean = false;
 
   constructor(config?: ServerConfig) {
     this._app = new Hono();
 
-    if (config)
-      this.configure(config);
-
-    this._app.use(defaultCors);
+    this.configure(config || { name: "default" });
   }
 
   public get app() {
@@ -26,35 +19,20 @@ export class BackendServer {
   }
 
   public configure(config: ServerConfig) {
-    this._config = config;
-
     // Base path
     if (config.basePath) this._app = this._app.basePath(config.basePath);
 
     // Database binding
-    this._app.use(async (ctx, next) => {
-      if (!BaseModelClass.db)
-        BaseModelClass.db = ctx.env.DB_PROXY;
+    this._app.use(bindDatabase);
 
-      await next();
-    });
+    // CORS
+    this._app.use(defaultCors);
 
     // Error handling
-    this._app.onError((err, ctx) => {
-      if (err instanceof APIError) {
-        ctx.status(err.statusCode || 500);
+    this._app.onError(defaultErrorHandler);
 
-        return ctx.json({
-          error_code: err.name,
-          error: err.message,
-          data: err.data
-        });
-      }
-
-      console.error(err);
-      ctx.status(500);
-      return ctx.json({ error: err.message });
-    });
+    // Standard middleware
+    if (config.standardMiddleware) this._app.use(...config.standardMiddleware);
 
     this.configured = true;
     return this;
